Refactor Balance component. Moving columns generator to helper class
