+++ src/routes/safe/component/Balances/Tokens.jsx
@@ -1,0 +1,10 @@
+// @flow
+import * as React from 'react'
+
+const Tokens = () => (
+  <React.Fragment>
+    Manage Tokens
+  </React.Fragment>
+)
+
+export default Tokens
